Added CStr and some other things to precompiled.h, to reduce build times. Altered CStr/UniDoubler's macros, to avoid polluting/conflicting/etc. Fixed headers in non-PCH builds. Used the _d version of vorbisfile*.lib in Debug builds.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2366 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/EntityStateProcessing.cpp
@@ -470,7 +470,7 @@ bool CEntity::processContactActionNoPathing( CEntityOrder* current, size_t times
 
 	// Load time needs to be animation->m_ActionPos2 ms after the start of the animation.
 
-	m_fsm_anipos2 = m_fsm_anipos + ( action->m_Speed * m_fsm_animation->m_ActionPos2 * 2 );
+	m_fsm_anipos2 = m_fsm_anipos + (size_t)( action->m_Speed * m_fsm_animation->m_ActionPos2 * 2 );
 	if( m_fsm_anipos2 < 0 )
 	{
 		// Load now.
